refactor(#87): formatNumber shared로 이동
